method-registration/requirements.txt
# Method Registration System Dependencies

# Configuration parsing
pyyaml>=6.0

# Database
psycopg2-binary>=2.9.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.90.0
pytest-asyncio>=0.21.0
testcontainers>=3.7.0
